address review: centralize providerLabel, fix oauthError scope, reuse types, fix backend multi-auth
Review comments addressed:
- Centralize providerLabel into shared lib/providerLabels.ts (was duplicated in 3 files)
- Only show oauthError when connector actually has OAuth credentials (hasOAuth check)
- Reuse exported OAuthConnection type from useOAuthConnections hook
- Fix GetRequiredCredentialByActionType to deterministically prefer oauth2
- Exclude oauth2 services from static credential resolution
- Add OAuth → static credential fallback for dual-auth connectors
