+++ frontend/src/pages/settings/ConnectedAccountsSection.tsx
@@ -22,15 +22,7 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
-
-const PROVIDER_LABELS: Record<string, string> = {
-  google: "Google",
-  microsoft: "Microsoft",
-};
-
-function providerLabel(id: string): string {
-  return PROVIDER_LABELS[id] ?? id.charAt(0).toUpperCase() + id.slice(1);
-}
+import { providerLabel } from "@/lib/providerLabels";
 
 function statusBadge(status: string) {
   switch (status) {
